backend/scripts/barchart_scraper/requirements.txt
# Barchart Scraper Dependencies

# Browser automation (proven working from backend/scraper.py)
playwright==1.40.0

# Google Sheets API
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Environment variables
python-dotenv==1.0.1
